model-space/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
llama-cpp-python==0.3.7
python-dotenv==1.0.1